fix connection request import on bus without geom
